requirements.txt
pandas 
numpy
scikit-learn
os
pickle
xgboost
streamlit
joblib
sklearn-pandas
category-encoders

